requirements.txt
google-api-python-client==2.110.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
Pillow==10.4.0
plyer==2.1.0
schedule==1.2.0
python-dotenv==1.0.0
